Move more code to NodeInlines.h https://bugs.webkit.org/show_bug.cgi?id=292243
Reviewed by Anne van Kesteren.

Moved more inline functions from Node.h to NodeInlines.h

* Source/WebCore/dom/Attr.cpp:
* Source/WebCore/dom/ContainerNode.cpp:
* Source/WebCore/dom/ContainerNodeAlgorithms.cpp:
* Source/WebCore/dom/Document.cpp:
* Source/WebCore/dom/Node.h:
(WebCore::Node::relaxAdoptionRequirement): Moved to NodeInlines.h.
(WebCore::Node::hasOneRef const): Ditto.
(WebCore::addSubresourceURL): Ditto.
(WebCore::Node::parentNodeGuaranteedHostFree const): Ditto.
(WebCore::Node::clearStyleFlags): Ditto.
(WebCore::Node::clearChildNeedsStyleRecalc): Ditto.
(WebCore::Node::setHasValidStyle): Ditto.
(WebCore::Node::setTreeScopeRecursively): Ditto.
(WebCore::Node::traverseToRootNodeInternal): Ditto.
* Source/WebCore/dom/NodeInlines.h:
(WebCore::Node::hasOneRef const): Moved from Node.h.
(WebCore::Node::clearStyleFlags): Ditto.
(WebCore::Node::clearChildNeedsStyleRecalc): Ditto.
(WebCore::Node::setHasValidStyle): Ditto.
(WebCore::Node::setTreeScopeRecursively): Ditto.
(WebCore::Node::parentNodeGuaranteedHostFree const): Ditto.
(WebCore::Node::traverseToRootNodeInternal): Ditto.
(WebCore::Node::relaxAdoptionRequirement): Ditto.
(WebCore::addSubresourceURL): Ditto.
* Source/WebCore/dom/ProcessingInstruction.cpp:
* Source/WebCore/html/HTMLBodyElement.cpp:
* Source/WebCore/html/HTMLEmbedElement.cpp:
* Source/WebCore/html/HTMLImageElement.cpp:
* Source/WebCore/html/HTMLInputElement.cpp:
* Source/WebCore/html/HTMLLinkElement.cpp:
* Source/WebCore/html/HTMLObjectElement.cpp:
* Source/WebCore/html/HTMLScriptElement.cpp:
* Source/WebCore/html/HTMLTableCellElement.cpp:
* Source/WebCore/html/HTMLTableElement.cpp:
* Source/WebCore/html/HTMLTemplateElement.cpp:
* Source/WebCore/html/track/TextTrackCue.cpp:
* Source/WebCore/page/DOMSelection.cpp:
* Source/WebCore/style/StyleTreeResolver.cpp:
* Source/WebCore/svg/SVGCursorElement.cpp:
* Source/WebCore/svg/SVGFEImageElement.cpp:
* Source/WebCore/svg/SVGImageElement.cpp:
* Source/WebCore/svg/SVGScriptElement.cpp:

Canonical link: https://commits.webkit.org/294302@main
