+++ Source/WebCore/html/HTMLBodyElement.cpp
@@ -35,6 +35,7 @@
 #include "JSHTMLBodyElement.h"
 #include "LocalDOMWindow.h"
 #include "MutableStyleProperties.h"
+#include "NodeInlines.h"
 #include "NodeName.h"
 #include "ResourceLoaderOptions.h"
 #include <wtf/NeverDestroyed.h>
