+++ Source/WebCore/html/HTMLTableElement.cpp
@@ -38,6 +38,7 @@
 #include "HTMLTableRowsCollection.h"
 #include "HTMLTableSectionElement.h"
 #include "MutableStyleProperties.h"
+#include "NodeInlines.h"
 #include "NodeName.h"
 #include "NodeRareData.h"
 #include "RenderTable.h"
