+++ Source/WebCore/style/StyleTreeResolver.cpp
@@ -45,6 +45,7 @@
 #include "LoaderStrategy.h"
 #include "LocalFrame.h"
 #include "MatchResultCache.h"
+#include "NodeInlines.h"
 #include "NodeRenderStyle.h"
 #include "Page.h"
 #include "PlatformStrategies.h"
